+++ docs/release-notes.rst
@@ -15,7 +15,9 @@ New API
 
 - Support of ``--config`` on the command line, to indicate a configuration file
 - Support of ``--disk-cache`` and ``--max-disk-cache-size`` on the command line
-
+- Support of ``webpage.settings.userName``, ``webpage.settings.password`` and ``webpage.settings.maxAuthAttemps``
+- New callback ``webpage.onAuthPrompt`` (not compatible with PhantomJS)
+- New method ``slimer.clearHttpAuth()`` to clear http authentication from the network cache.
 
 Fixed bugs
 ----------
@@ -31,7 +33,7 @@ Missing APIS in SlimerJS 0.9
 Here are the PhantomJS APIs that are missing in SlimerJS 0.9 but planed in future releases.
 
 - most of options for the command line are not supported
-- no support of all settings on the webpage object, so no support of HTTP authentication...
+- some settings on the webpage object are not supported yet
 - no support of Ghost Driver (Selenium web driver)
 
 You can read the `compatibility table <https://github.com/laurentj/slimerjs/blob/master/API_COMPAT.md>`_ to know details.